# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Modelo de API para transações da Saque e Pague. O parceiro interessado\
    \ deve implementar as operações desejadas para que a SeP consiga se comunicar\
    \ com o mesmo."
  version: "1.19.0"
  title: "Saque e Pague Transacional"
  contact:
    name: "Time Saque e Pague"
    url: "http://www.saqueepague.com.br"
host: "api-saqueepague.sensedia.com"
basePath: "/saqueepague/SepTransaction/1.19.0"
...
schemes:
- "https"
paths:
...
  /consultaPagamento:
    post:
      tags:
      - "Pagamento com Cartão"
      - "Pagamento sem Cartão"
      description: "Operação de consulta de informações do boleto a ser pago em dinheiro\
        \ ou débito em conta."
      operationId: "consultaPagamentoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição para consulta de pagamento de boleto com cartão."
        required: true
        schema:
          $ref: "#/definitions/consultaPagamentoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/consultaPagamentoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "PagamentoComCarto"
...
  /extrato:
    post:
      tags:
      - "Extrato com Cartão"
      description: "Operação de extrato."
      operationId: "extratoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de operação de extrato."
        required: true
        schema:
          $ref: "#/definitions/extratoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/extratoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "ExtratoComCarto"
  /extratoConf:
    post:
      tags:
      - "Extrato com Cartão"
      description: "Confirmação de operação de extrato."
      operationId: "extratoConfPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de confirmação de operação de extrato."
        required: true
        schema:
          $ref: "#/definitions/transacConf"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "ExtratoComCarto"
  /identificacao:
    post:
      tags:
      - "Saque com Cartão"
      - "Saldo com Cartão"
      - "Extrato com Cartão"
      - "Depósito com Cartão"
      - "Empréstimo com Cartão"
      - "Pagamento com Cartão"
      - "Transferência com Cartão"
      description: "Operação de identificação do cliente e informações adicionais\
        \ de sua conta."
      operationId: "identificacaoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de identificação."
        required: true
        schema:
          $ref: "#/definitions/identificacaoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/identificacaoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "SaqueComCarto"
  /pagamento:
    post:
      tags:
      - "Pagamento com Cartão"
      - "Pagamento sem Cartão"
      description: "Operação de pagamento de boleto de cobrança com cartão."
      operationId: "pagamentoPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição para pagamento de boleto com cartão."
        required: true
        schema:
          $ref: "#/definitions/pagamentoReq"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
          schema:
            $ref: "#/definitions/pagamentoResp"
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "PagamentoComCarto"
  /pagamentoConf:
    post:
      tags:
      - "Pagamento com Cartão"
      - "Pagamento sem Cartão"
      description: "Confirmação de operação de pagamento de boleto de cobrança."
      operationId: "pagamentoConfPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "authenticationType"
        in: "header"
        description: "Tipo de autenticação requerida."
        required: true
        type: "string"
        x-example: "auth2"
      - name: "clientId"
        in: "header"
        description: "Identificação do cliente."
        required: true
        type: "string"
        x-example: "prd-client"
      - name: "token"
        in: "header"
        description: "Chave para validação do acesso ao serviço."
        required: true
        type: "string"
        x-example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjYsInByb2ZpbGUiOltdLCJpYXQiOjE0OTkwOsY4NDV9.uR5WH3c2ZztGmZVCfcJPu-1ix1w-j7shJw_X_ZlFfAo"
      - in: "body"
        name: "body"
        description: "Requisição de confirmação de operação de pagamento."
        required: true
        schema:
          $ref: "#/definitions/transacConf"
      responses:
        200:
          description: "Retorno com sucesso ou com Erro de negócio."
        400:
          description: "Bad Request ou Erro interno ao qual inviabilizou uma resposta."
        401:
          description: "Acesso não autorizado."
      x-swagger-router-controller: "PagamentoComCarto"
...
definitions:
...
  cartao:
    type: "object"
    properties:
      modoEntrada:
        type: "string"
        description: "Modo de entrada de leitura do cartão. (000 = sem cartão, 050\
          \ = chip, 800 = chip, se falhar realiza leitura da trilha, 900 = trilha\
          \ completa)."
        enum:
        - "000"
        - "050"
        - "800"
        - "900"
      pan:
        type: "string"
        example: "zhXTyHl5nf2ckysYXXjAgZS5fzS7ybC5xePnWeLEP//bw=="
        description: "PAN/Número do cartão utilizado (dado criptografado)."
        pattern: "^.{10,}$"
      trilha2:
        type: "string"
        example: "mMIrkIrf2pSHccWuaROxoIS+NGcTOWcTWXUtjniTa1BRzPziiE8pWnknAxW9HIgoKdc/Fli2KZwNA=="
        description: "Trilha 2 do cartão utilizado (dado criptografado)."
        pattern: "^.{10,}$"
...
  cripto:
    type: "object"
    properties:
      criptogramaChip:
        type: "string"
        description: "Em caso de cartão com chip, contém o criptograma dele."
      hash:
        type: "string"
        description: "Hash de todos os campos que compõem a transação."
      idPositiva:
        type: "string"
        example: "q0WKgCM+wbXjzlY="
        description: "Identificação positiva informada pelo usuário, caso utilizada,\
          \ com no mínimo 8 dígitos corresponde às posições (botões) em cada tela\
          \ apresentada. Caso a senha tenha menos de oito sílabas, são adicionados\
          \ zeros para as telas não existentes (dado criptografado)."
        pattern: "^.{8,}$"
      respostaCliente:
        $ref: "#/definitions/cripto_respostaCliente"
      senha:
        type: "string"
        example: "q0WKgCM+wbXjzlY="
        description: "Senha numérica criptografada."
      tamanhoSenha:
        type: "string"
        example: "15"
        description: "Tamanho da senha numérica."
        pattern: "^\\d+$"
      templateBiometria:
        type: "string"
        description: "Em caso de biometria, contém o template gerado."
      token:
        type: "string"
        example: "02C47DF604EB43B1"
        description: "Em caso de utilização de token para autenticação, contém o token\
          \ informado pelo usuário."
    description: "Informações relacionadas a criptografia e senhas da requisição.\
      \ Varia dependendo da necessidade de cada transação."
  criptoResp:
    type: "object"
    properties:
      hash:
        type: "string"
        description: "Hash de todos os campos que compõem a transação."
    example:
      hash: "hash"
...
  extratoReq:
    type: "object"
    required:
    - "InfExtrato"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfExtrato:
        $ref: "#/definitions/infExtratoReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  extratoResp:
    type: "object"
    required:
    - "InfExtrato"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfExtrato:
        $ref: "#/definitions/infExtratoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfExtrato:
        recibo: "@                EXTRATO DE CONTA                @              \
          \                                  @ CLIENTE: EUGENIO SCHMITT COELHO   \
          \             @ AGENCIA: 0150 CONTA: 03.013689.0-1             @ SALDO DA\
          \ CONTA                                 @ SALDO DEVEDOR...............R$\
          \          450,35-@ TOTAL DEVEDOR...............R$          450,35-@ LIMITE\
          \ DA CONTA.............R$          100,00 @ LIMITE DA CONTA DISPONIVEL..R$\
          \          350,35-@--------- MOVIMENTOS DA CONTA CORRENTE ---------@@  \
          \  SALDO ANT EM 18/01/2019               93,56-@    MOVIMENTOS JAN/2019\
          \                         @    COMPRAS           201218              85,68-@\
          \    COMPRAS           201218             139,14-@    COMPRAS          \
          \ 211218              20,70-@    COMPRAS           211218              39,00-@\
          \    SALDO NA DATA                        378,08-@"
        qtdPaginas: "000010"
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
  identificacaoReq:
    type: "object"
    required:
    - "Cartao"
    - "Cripto"
    - "InfIdentificacao"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cartao:
        $ref: "#/definitions/cartao"
      Cripto:
        $ref: "#/definitions/cripto"
      InfIdentificacao:
        $ref: "#/definitions/infIdentificacaoReq"
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Terminal:
        $ref: "#/definitions/terminal"
  identificacaoResp:
    type: "object"
    required:
    - "InfIdentificacao"
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfIdentificacao:
        $ref: "#/definitions/infIdentificacaoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
    example:
      InfIdentificacao:
        pedirToken: "01"
        frase: "Informe o token gerado em seu app:"
        numAgencia: "4029"
        perguntas:
        - pergunta: "Qual é o dia de seu aniversário?"
          tamResposta: "2"
          id: "1"
        - pergunta: "Qual é o dia de seu aniversário?"
          tamResposta: "2"
          id: "1"
        idPositiva:
        - pos8: "4E-8G-1P"
          idTela: "1"
          pos4: "P1-G8-E4"
          pos5: "REG SOS"
          pos6: "EN/UQ/AM"
          pos7: "69 05 41"
          pos1: "SOS GER"
          pos2: "MA/QU/NE"
          pos3: "14 50 96"
        - pos8: "4E-8G-1P"
          idTela: "1"
          pos4: "P1-G8-E4"
          pos5: "REG SOS"
          pos6: "EN/UQ/AM"
          pos7: "69 05 41"
          pos1: "SOS GER"
          pos2: "MA/QU/NE"
          pos3: "14 50 96"
        numConta: "0082348296"
        transacValidas: "068101068010"
        saldo: "000000005000"
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Cripto:
        hash: "hash"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
...
  infConsultaPagamentoReq:
    type: "object"
    properties:
      cpf:
        type: "string"
        example: "01234567890"
        description: "Número do CPF ou CNPJ (11 dígitos ou 14 dígitos respectivamente)."
        pattern: "^\\d{14}|\\d{11}$"
      codBarras:
        $ref: "#/definitions/infConsultaPagamentoReq_codBarras"
  infConsultaPagamentoResp:
    type: "object"
    required:
    - "cedente"
    - "dataPagamento"
    - "desconto"
    - "valor"
    properties:
      valor:
        type: "string"
        example: "000000005000"
        description: "Valor do pagamento (12 dígitos, incluindo centavos)."
        pattern: "^\\d{12}$"
      cedente:
        type: "string"
        example: "Saque e Pague"
        description: "Cedente (beneficiário)."
      dataPagamento:
        type: "string"
        example: "20181122"
        description: "Data de pagamento atualizada pelo parceiro (AAAAMMDD)."
        pattern: "^\\d{8}$"
      mensagem:
        type: "string"
        example: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em tela."
        description: "Mensagem informativa quanto à alteração de valores e/ou data\
          \ de pagamento pelo paceiro."
      desconto:
        type: "string"
        example: "000000000499"
        description: "Valor de desconto (12 dígitos, incluindo centavos)."
        pattern: "^\\d{12}$"
      codBarras:
        $ref: "#/definitions/infConsultaPagamentoResp_codBarras"
      cip:
        $ref: "#/definitions/infConsultaPagamentoResp_cip"
      aceitaPagamentoParcial:
        type: "string"
        example: "01"
        description: "Aceita pagamento parcial (00 = não, 00 = sim)."
        enum:
        - "00"
        - "01"
      valorMinimo:
        type: "string"
        example: "000000000500"
        description: "Valor mínimo do pagamento (12 dígitos, incluindo centavos)."
        pattern: "^\\d{12}$"
    description: "Informações da resposta da requisição de consulta de pagamento."
    example:
      aceitaPagamentoParcial: "01"
      dataPagamento: "20181122"
      desconto: "000000000499"
      mensagem: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em tela."
      valor: "000000005000"
      cip:
        valorMaximoPermitidoPagamento: "000000050000"
        valorMinimoPermitidoPagamento: "000000000500"
        valorJurosCalculado: "000000000499"
        valorAbatimento: "000000000499"
        tipoValorAceito: "3"
        valorMultaCalculadada: "000000000499"
      valorMinimo: "000000000500"
      codBarras:
        codigo: "12345678912346579812345678912345678913456789123456789"
        dataVencimento: "20181122"
      cedente: "Saque e Pague"
...
  infTransacaoReq:
    type: "object"
    required:
    - "cdProc"
    - "codOperadora"
    - "dataHora"
    - "dataLocal"
    - "horaLocal"
    - "nsu"
    properties:
      cdProc:
        type: "string"
        example: "029100"
        description: "Código de processamento da transação, gerado pela Saque e Pague.\
          \ Verificar código específico de cada transação com a Saque e Pague (6 dí\
          gitos)."
        pattern: "^\\d{6}$"
      codMoeda:
        type: "string"
        example: "986"
        description: "Código na tabela de moedas da ISO-4217 (986 = real, 3 dígitos)."
        pattern: "^\\d{3}$"
      codOperadora:
        type: "string"
        example: "00000000914"
        description: "Código definido pela Saque e Pague que identifica o autorizador\
          \ (11 dígitos)."
        pattern: "^\\d{11}$"
      dataContabil:
        type: "string"
        example: "1122"
        description: "Data contábil da transação (MMDD)."
        pattern: "^\\d{4}$"
      dataHora:
        type: "string"
        example: "1122151032"
        description: "Data e hora da transação (MMDDhhmmss)."
        pattern: "^\\d{10}$"
      dataLocal:
        type: "string"
        example: "1122"
        description: "Data local (MMDD)."
        pattern: "^\\d{4}$"
      horaLocal:
        type: "string"
        example: "151032"
        description: "Hora local (hhmmss)."
        pattern: "^\\d{6}$"
      nsu:
        type: "string"
        example: "000080247206"
        description: "Identificador da transação gerado pela Saque e Pague (de 9 dí\
          gitos a 12 dígitos)."
        pattern: "^\\d{12}$"
      valor:
        type: "string"
        example: "000000005000"
        description: "Valor da transação (12 dígitos, incluindo centavos)."
        pattern: "^\\d{12}$"
    description: "Informações genéricas sobre a requisição de transação."
  infTransacaoResp:
    type: "object"
    required:
    - "cdProc"
    - "codOperadora"
    - "dataHora"
    - "dataLocal"
    - "horaLocal"
    - "nsu"
    - "nsuResposta"
    properties:
      cdProc:
        type: "string"
        example: "029100"
        description: "Igual à requisição."
        pattern: "^\\d{6}$"
      codMoeda:
        type: "string"
        example: "986"
        description: "Igual à requisição com exceção da transação de consultaSaque."
        pattern: "^\\d{3}$"
      codOperadora:
        type: "string"
        example: "00000000914"
        description: "Igual à requisição."
        pattern: "^\\d{11}$"
      dataHora:
        type: "string"
        example: "1122151032"
        description: "Data e hora da transação (MMDDhhmmss)."
        pattern: "^\\d{10}$"
      dataLocal:
        type: "string"
        example: "1122"
        description: "Igual à requisição."
        pattern: "^\\d{4}$"
      errorMessage:
        type: "string"
        example: ""
        description: "Caso ocorra erro, este campo preenchido identifica que um erro\
          \ ocorreu e seu valor."
      horaLocal:
        type: "string"
        example: "151032"
        description: "Igual à requisição."
        pattern: "^\\d{6}$"
      mensagemCliente:
        type: "string"
        example: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em tela."
        description: "Mensagem para ser mostrada na tela para o cliente, em caso de\
          \ erro (objeto errorMessage deve estar preenchido para identificação de\
          \ um erro)."
      nsu:
        type: "string"
        example: "000080247206"
        description: "Igual à requisição."
        pattern: "^\\d{12}$"
      nsuResposta:
        type: "string"
        example: "820"
        description: "Identificador da resposta da transação, gerado pelo autorizador."
        pattern: "^\\d+$"
      valor:
        type: "string"
        example: "5000"
        description: "Igual à requisição com exceção da transação de consultaSaldo\
          \ e consultaSaque."
        pattern: "^\\d+$"
    description: "Informações sobre a resposta da requisição de transação."
    example:
      cdProc: "029100"
      mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
        \ tela."
      nsu: "000080247206"
      codMoeda: "986"
      codOperadora: "00000000914"
      dataLocal: "1122"
      errorMessage: ""
      valor: "5000"
      horaLocal: "151032"
      nsuResposta: "820"
      dataHora: "1122151032"
...
  pagamentoReq:
    type: "object"
    required:
    - "Cartao"
    - "InfPagamento"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Cartao:
        $ref: "#/definitions/cartao"
      Terminal:
        $ref: "#/definitions/terminal"
      Cripto:
        $ref: "#/definitions/cripto"
      InfPagamento:
        $ref: "#/definitions/infPagamentoReq"
  pagamentoResp:
    type: "object"
    required:
    - "InfPagamentoResp"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfPagamentoResp:
        $ref: "#/definitions/infPagamentoResp"
    example:
      InfPagamentoResp:
        recibo: "@            COMPROVANTE DE PAGAMENTO            @              \
          \                                  @ CEDENTE: 008 BANCO SAQUE E PAGUE  \
          \             @ VALOR..: 263,85                                @ DATA DO\
          \ PAGAMENTO.: 29/01/2019                 @ CODIGO DE BARRAS            \
          \                   @@   84600000000-3 25850072001-1                 @ \
          \  10357244475-5 01192190121-8                  @"
        qtdeViasComprovante: "1"
        dataPagamento: "20181122"
        desconto: "000000000499"
        mensagem: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em tela."
        valor: "000000005000"
        cedente: "Saque e Pague"
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
      Cripto:
        hash: "hash"
  consultaPagamentoReq:
    type: "object"
    required:
    - "Cartao"
    - "InfConsultaPagamento"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfTransacao:
        $ref: "#/definitions/infTransacaoReq"
      Cartao:
        $ref: "#/definitions/cartao"
      Terminal:
        $ref: "#/definitions/terminal"
      InfConsultaPagamento:
        $ref: "#/definitions/infConsultaPagamentoReq"
  consultaPagamentoResp:
    type: "object"
    required:
    - "InfConsultaPagamento"
    - "InfTransacao"
    - "Terminal"
    properties:
      InfTransacao:
        $ref: "#/definitions/infTransacaoResp"
      Terminal:
        $ref: "#/definitions/terminal"
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfConsultaPagamento:
        $ref: "#/definitions/infConsultaPagamentoResp"
    example:
      InfTransacao:
        cdProc: "029100"
        mensagemCliente: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em\
          \ tela."
        nsu: "000080247206"
        codMoeda: "986"
        codOperadora: "00000000914"
        dataLocal: "1122"
        errorMessage: ""
        valor: "5000"
        horaLocal: "151032"
        nsuResposta: "820"
        dataHora: "1122151032"
      Terminal:
        codEstab: "000000000742673"
        tipo: "008"
        id: "05100004"
      Cripto:
        hash: "hash"
      InfConsultaPagamento:
        aceitaPagamentoParcial: "01"
        dataPagamento: "20181122"
        desconto: "000000000499"
        mensagem: "Esta &eacute; uma mensagem para exibi&ccedil;&atilde;o em tela."
        valor: "000000005000"
        cip:
          valorMaximoPermitidoPagamento: "000000050000"
          valorMinimoPermitidoPagamento: "000000000500"
          valorJurosCalculado: "000000000499"
          valorAbatimento: "000000000499"
          tipoValorAceito: "3"
          valorMultaCalculadada: "000000000499"
        valorMinimo: "000000000500"
        codBarras:
          codigo: "12345678912346579812345678912345678913456789123456789"
          dataVencimento: "20181122"
        cedente: "Saque e Pague"
...
  terminal:
    type: "object"
    required:
    - "codEstab"
    - "id"
    - "tipo"
    properties:
      codEstab:
        type: "string"
        example: "000000000742673"
        description: "Código do estabelecimento onde está localizado o terminal gerado\
          \ pela Saque e Pague (15 dígitos)."
        pattern: "^\\d{15}$"
      id:
        type: "string"
        example: "05100004"
        description: "Identificador do terminal gerado pela Saque e Pague (8 dígitos)."
        pattern: "^\\d{8}$"
      tipo:
        type: "string"
        example: "008"
        description: "Tipo de terminal o qual a operação é gerada (3 dígitos, 008\
          \ = ATM)."
        pattern: "^\\d{3}$"
    description: "Informações sobre o terminal vigente."
    example:
      codEstab: "000000000742673"
      tipo: "008"
      id: "05100004"
  transacConf:
    type: "object"
    required:
    - "InfTransacao"
    - "Terminal"
    properties:
      Cripto:
        $ref: "#/definitions/criptoResp"
      InfTransacao:
        $ref: "#/definitions/infTransacaoConf"
      Terminal:
        $ref: "#/definitions/terminal"
...
  infConsultaPagamentoReq_codBarras:
    properties:
      codigo:
        type: "string"
        example: "12345678912346579812345678912345678913456789123456789"
        description: "Código de barras do documento."
        pattern: "^\\d+$"
      modoEntrada:
        type: "string"
        example: "0"
        description: "Forma de entrada do código de barras (0 = escaneado, 1 = digitado)."
        enum:
        - "0"
        - "1"
...